% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice.R
\name{col_means}
\alias{col_means}
\title{Calculate Column Means}
\usage{
col_means(df)
}
\arguments{
\item{df}{A data frame containing numeric columns.}
}
\value{
A numeric vector of column means.
}
\description{
This function takes a data frame and returns a vector of column means.
}
\examples{
df <- data.frame(col1 = c(1, 2, 3), col2 = c(4, 5, 6), col3 = c(7, 8, NA))
col_means(df)

}
