% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice.R
\name{count_na}
\alias{count_na}
\title{Count NA Values}
\usage{
count_na(vec)
}
\arguments{
\item{vec}{A numeric vector.}
}
\value{
An integer count of NA values in the vector.
}
\description{
This function counts the number of NA values in a given vector.
}
\examples{
vec  <- c(1, NA, 3, NA, 5)
count_na(vec)

}
