% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice.R
\name{col_means}
\alias{col_means}
\title{Title takes as input a data frame and returns a vector of column means}
\usage{
col_means(df)
}
\arguments{
\item{df}{inputs a data frame}
}
\value{
returns a vector of column means
}
\description{
Title takes as input a data frame and returns a vector of column means
}
\examples{
sample_df <- data.frame(
A = c(1, 2, 3),
B = c(4, 5, 6),
C = c(7, 8, 9))
col_means(sample_df)
}
