% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice.R
\name{count_na}
\alias{count_na}
\title{Title count_na() returns how many NA’s there are in a vector}
\usage{
count_na(vec)
}
\arguments{
\item{vec}{inputs a vector}
}
\value{
the count of NA's in a vector
}
\description{
Title count_na() returns how many NA’s there are in a vector
}
\examples{
count_na(c(1, NA, 3, NA, 5, NA))
}
